+++ ChangeLog.md
@@ -4,8 +4,11 @@ ___
 
 ## Unreleased
 
+
+- NEW : support supplier invoices - *2021-06-16* - 1.15.0
 - FIX : Add contact to ticket on creation - *2021-06-20* - 1.14.4
 - FIX : V14 compatibility : Edit the descriptor: $this->family - *2021-06-10* - 1.14.4
+
 - NEW : Option to set favicon
 
 ## RELEASE 1.13 - 2021-03-19